Layout_TestCase/http/tests/xmlhttprequest/origin-whitelisting-https-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
Tests that origin whitelisting for https does not match http URLs.

Testing: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (sync)
PASS: Error: NETWORK_ERR: XMLHttpRequest Exception 101
Testing: http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/get.txt (async)
PASS: 0
